requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
openai==1.3.0
celery==5.3.4
redis==5.0.1
boto3==1.28.64
python-jose==3.3.0
python-multipart==0.0.6
prometheus-client==0.17.1
structlog==23.2.0
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
